create getDates
